adjust file structure: put data access function in new file
