+++ chord/data_access.go
@@ -1,0 +1,167 @@
+package chord
+
+import (
+	"fmt"
+	"math/big"
+
+	"github.com/sirupsen/logrus"
+)
+
+func (node *Node) putData(pair Pair) error {
+	node.dataLock.Lock()
+	node.data[pair.Key] = pair.Value
+	node.dataLock.Unlock()
+	var suc SingleNode
+	err := node.get_successor(&suc)
+	if err != nil {
+		logrus.Warnf("Warning! <func putData()> %v", err)
+		return err
+	}
+	err = node.RemoteCall("tcp", suc.Addr, "RPC_Node.PutBackupData", pair, &struct{}{})
+	if err != nil {
+		logrus.Errorf("Error <func putData()> node [%s] call [%s] method [RPC_Node.PutBackupData] error: %v", node.getPort(), suc.getPort(), err)
+		return err
+	}
+	return nil
+}
+
+func (node *Node) putDataList(dataList map[string]string) error {
+	node.dataLock.Lock()
+	for key, value := range dataList {
+		node.data[key] = value
+	}
+	node.dataLock.Unlock()
+	var suc SingleNode
+	err := node.get_successor(&suc)
+	if err != nil {
+		logrus.Warnf("Warning! <func putDataList()> %v", err)
+		return err
+	}
+	err = node.RemoteCall("tcp", suc.Addr, "RPC_Node.PutBackupDataList", dataList, &struct{}{})
+	if err != nil {
+		logrus.Errorf("Error <func putDataList()> node [%s] call [%s] method [RPC_Node.PutBackupDataList] error: %v", node.getPort(), suc.getPort(), err)
+		return err
+	}
+	return nil
+}
+
+func (node *Node) putBackupData(pair Pair) error {
+	node.backupDataLock.Lock()
+	node.backupData[pair.Key] = pair.Value
+	node.backupDataLock.Unlock()
+	return nil
+}
+
+func (node *Node) putBackupDataList(backupDataList map[string]string) error {
+	node.backupDataLock.Lock()
+	for key, value := range backupDataList {
+		node.backupData[key] = value
+	}
+	node.backupDataLock.Unlock()
+	return nil
+}
+
+func (node *Node) getData(key string, value *string) error {
+	node.dataLock.RLock()
+	*value = node.data[key]
+	node.dataLock.RUnlock()
+	return nil
+}
+
+func (node *Node) getDataList(dataList *map[string]string) error {
+	node.dataLock.RLock()
+	*dataList = node.data
+	node.dataLock.RUnlock()
+	return nil
+}
+
+func (node *Node) getBackupDataList(backupDataList *map[string]string) error {
+	node.backupDataLock.Lock()
+	*backupDataList = node.backupData
+	node.backupDataLock.Unlock()
+	return nil
+}
+
+func (node *Node) deleteData(key string) error {
+	node.dataLock.Lock()
+	_, ok := node.data[key]
+	delete(node.data, key)
+	node.dataLock.Unlock()
+	if !ok {
+		return fmt.Errorf("no key [%s] on node [%s]'s data", key, node.getPort())
+	}
+	var suc SingleNode
+	err := node.get_successor(&suc)
+	if err != nil {
+		logrus.Warnf("Warning! <func deletaData()> %v", err)
+		return err
+	}
+	err = node.RemoteCall("tcp", suc.Addr, "RPC_Node.DeleteBackupData", key, &struct{}{})
+	if err != nil {
+		logrus.Errorf("Error <func deleteData()> node [%s] call [%s] method [RPC_Node.DeleteBackupData] error: %v", node.getPort(), suc.getPort(), err)
+		return err
+	}
+	return nil
+}
+
+func (node *Node) deleteDataList(dataList map[string]string) error {
+	node.dataLock.Lock()
+	for key := range dataList {
+		delete(node.data, key)
+	}
+	node.dataLock.Unlock()
+	var suc SingleNode
+	err := node.get_successor(&suc)
+	if err != nil {
+		logrus.Warnf("Warning! <func deleteDataList()> %v", err)
+		return err
+	}
+	err = node.RemoteCall("tcp", suc.Addr, "RPC_Node.DeleteBackupDataList", dataList, &struct{}{})
+	if err != nil {
+		logrus.Errorf("Error <func deleteDataList()> node [%s] call [%s] method [RPC_Node.DeleteBackupDataList] error: %v", node.getPort(), suc.getPort(), err)
+		return err
+	}
+	return nil
+}
+
+func (node *Node) deleteBackupData(key string) error {
+	node.backupDataLock.Lock()
+	_, ok := node.backupData[key]
+	delete(node.backupData, key)
+	node.backupDataLock.Unlock()
+	if !ok {
+		return fmt.Errorf("no key [%s] on node [%s]'s backupData", key, node.getPort())
+	}
+	return nil
+}
+
+func (node *Node) deleteBackupDataList(backupDataList map[string]string) error {
+	node.backupDataLock.Lock()
+	for key := range backupDataList {
+		delete(node.backupData, key)
+	}
+	node.backupDataLock.Unlock()
+	return nil
+}
+
+func (node *Node) transferData(pre SingleNode, pre_data *map[string]string) error {
+	node.backupData = make(map[string]string)
+	var id *big.Int
+	node.dataLock.Lock()
+	for key, value := range node.data {
+		id = Hash(key)
+		if !(in_range(id, pre.ID, node.ID) || id.Cmp(node.ID) == 0) {
+			(*pre_data)[key] = value
+			node.backupDataLock.Lock()
+			node.backupData[key] = value
+			node.backupDataLock.Unlock()
+			delete(node.data, key)
+		}
+	}
+	node.dataLock.Unlock()
+	var suc SingleNode
+	node.get_successor(&suc)
+	node.RemoteCall("tcp", suc.Addr, "RPC_Node.DeleteBackupDataList", node.backupData, &struct{}{})
+	node.set_predecessor(&pre)
+	return nil
+}
